EDU-208 Adjust the styles and remove faulty room js
