+++ src/components/pages/room.less
@@ -1,0 +1,9 @@
+.Room {}
+
+.Room-membersCollapse {
+  margin-top: 16px;
+}
+
+.Room-invitationsCollapse {
+  margin-top: 16px;
+}
