+++ src/styles/main.less
@@ -46,6 +46,7 @@
 @import '../components/pages/import-batch-view.less';
 @import '../components/pages/browser-not-supported.less';
 @import '../components/pages/room-membership-confirmation.less';
+@import '../components/pages/room.less';
 @import '../components/link-popover.less';
 @import '../components/audio-player.less';
 @import '../components/color-picker.less';
